test: 로또 결과 확인 테스트 코드 이동
