+++ src/test/java/lotto/model/PurchasedLottosTest.java
@@ -1,7 +1,6 @@
 package lotto.model;
 
 import java.util.List;
-import java.util.Map;
 import lotto.util.FixedNumbersGenerator;
 import lotto.util.NumbersGenerator;
 import org.junit.jupiter.api.Assertions;
@@ -28,21 +27,6 @@ class PurchasedLottosTest {
 
         purchasedLottos.forEach(lotto -> Assertions.assertEquals("[1, 2, 3, 4, 5, 6]", lotto.toString()));
 
-    }
-
-    @Test
-    @DisplayName("로또 결과 확인 테스트")
-    void 로또_결과_확인_테스트() {
-        int purchaseCount = 3;
-        PurchasedLottos lottos = PurchasedLottos.purchase(purchaseCount, numbersGenerator);
-
-        Lotto winningNumbers = Lotto.of(List.of(1, 2, 3, 4, 5, 6));
-        int bonusNumber = 7;
-        WinningLotto winningLotto = WinningLotto.of(winningNumbers, bonusNumber);
-
-        Map<LottoRank, Integer> lottoResult = lottos.lottoResultFrom(winningLotto);
-
-        Assertions.assertEquals(3, lottoResult.get(LottoRank.FIRST));
     }
 
     @Test
